refactor(refs T30969): move DpDataTable to components root
