+++ src/components/DpDataTable/DpDataTable.vue
@@ -1,8 +1,8 @@
 <script>
-import { CleanHtml } from '../../../directives/CleanHtml/CleanHtml'
+import { CleanHtml } from '../../directives/CleanHtml/CleanHtml'
 import DomPurify from 'dompurify'
-import DpDraggable from '../../DpDraggable/DpDraggable'
-import DpLoading from '../../DpLoading/DpLoading'
+import DpDraggable from '../DpDraggable/DpDraggable'
+import DpLoading from '../DpLoading/DpLoading'
 import DpTableHeader from './DpTableHeader'
 import DpTableRow from './DpTableRow'
