+++ src/components/DpDataTable/DpDataTableExtended.vue
@@ -104,10 +104,10 @@ import dataTableSearch from './DataTableSearch'
 import DomPurify from 'dompurify'
 import DpDataTable from './DpDataTable'
 import DpSelectPageItemCount from './DpSelectPageItemCount'
-import DpSlidingPagination from '../../DpSlidingPagination/DpSlidingPagination'
-import DpStickyElement from '../shared/DpStickyElement'
-import { hasOwnProp } from '../../../utils'
-import { tableSelectAllItems } from '../../../mixins'
+import DpSlidingPagination from '../DpSlidingPagination/DpSlidingPagination'
+import DpStickyElement from '../core/shared/DpStickyElement'
+import { hasOwnProp } from '../../utils'
+import { tableSelectAllItems } from '../../mixins'
 
 export default {
   name: 'DpDataTableExtended',
