+++ src/components/DpDataTable/DpTableHeader.vue
@@ -62,7 +62,7 @@
 </template>
 
 <script>
-import DpIcon from '../../DpIcon/DpIcon'
+import DpIcon from '../DpIcon/DpIcon'
 import DpWrapTrigger from './DpWrapTrigger'
 import DpResizableColumn from './DpResizableColumn'
