+++ src/components/DpDataTable/DpWrapTrigger.vue
@@ -9,7 +9,7 @@
 </template>
 
 <script>
-import DpIcon from '../../DpIcon/DpIcon'
+import DpIcon from '../DpIcon/DpIcon'
 
 export default {
   name: 'DpWrapTrigger',
