+++ src/components/core/index.js
@@ -1,18 +1,13 @@
-import dataTableSearch from './DpDataTable/DataTableSearch'
 import DpAnonymizeText from './DpAnonymizeText'
 import DpBulkEditHeader from './DpBulkEditHeader'
 import DpChangeStateAtDate from './DpChangeStateAtDate'
-import DpColumnSelector from './DpDataTable/DpColumnSelector'
 import DpCopyPasteButton from './DpCopyPasteButton'
 import DpDashboardTaskCard from './DpDashboardTaskCard'
-import DpDataTable from './DpDataTable/DpDataTable'
-import DpDataTableExtended from './DpDataTable/DpDataTableExtended'
 import DpEditableList from './DpEditableList'
 import DpEditor from './DpEditor/DpEditor'
 import DpNotifyContainer from './notify/DpNotifyContainer'
 import DpNotifyMessage from './notify/DpNotifyMessage'
 import DpObscure from './DpObscure'
-import DpSelectPageItemCount from './DpDataTable/DpSelectPageItemCount'
 import DpStickyElement from './shared/DpStickyElement'
 import DpTableCard from './DpTableCardList/DpTableCard'
 import DpTableCardListHeader from './DpTableCardList/DpTableCardListHeader'
@@ -23,21 +18,16 @@ import DpVideoPlayer from './DpVideoPlayer'
 import { getFileIdsByHash } from './DpUpload/utils/GetFileIdsByHash'
 
 export {
-  dataTableSearch,
   DpAnonymizeText,
   DpBulkEditHeader,
   DpChangeStateAtDate,
-  DpColumnSelector,
   DpCopyPasteButton,
   DpDashboardTaskCard,
-  DpDataTable,
-  DpDataTableExtended,
   DpEditableList,
   DpEditor,
   DpNotifyContainer,
   DpNotifyMessage,
   DpObscure,
-  DpSelectPageItemCount,
   DpStickyElement,
   DpTableCard,
   DpTableCardListHeader,
