+++ src/components/index.js
@@ -1,3 +1,4 @@
+import dataTableSearch from './DpDataTable/DataTableSearch'
 import DpAccordion from './DpAccordion/DpAccordion'
 import DpAutocomplete from './DpAutocomplete/DpAutocomplete'
 import DpButton from './DpButton/DpButton'
@@ -6,7 +7,10 @@ import DpButtonRow from './DpButtonRow/DpButtonRow'
 import DpCard from './DpCard/DpCard'
 import DpCheckbox from './DpCheckbox/DpCheckbox'
 import DpCheckboxGroup from './DpCheckboxGroup/DpCheckboxGroup'
+import DpColumnSelector from './DpDataTable/DpColumnSelector'
 import DpContextualHelp from './DpContextualHelp/DpContextualHelp'
+import DpDataTable from './DpDataTable/DpDataTable'
+import DpDataTableExtended from './DpDataTable/DpDataTableExtended'
 import DpDatepicker from './DpDatepicker/DpDatepicker'
 import DpDateRangePicker from './DpDateRangePicker/DpDateRangePicker'
 import DpDatetimePicker from './DpDatetimePicker/DpDatetimePicker'
@@ -29,6 +33,7 @@ import DpRadio from './DpRadio/DpRadio'
 import DpResettableInput from './DpResettableInput/DpResettableInput'
 import DpSearchField from './DpSearchField/DpSearchField'
 import DpSelect from './DpSelect/DpSelect'
+import DpSelectPageItemCount from './DpDataTable/DpSelectPageItemCount'
 import DpSkeletonBox from './DpSkeletonBox/DpSkeletonBox'
 import DpSlidebar from './DpSlidebar/DpSlidebar'
 import DpSlidingPagination from './DpSlidingPagination/DpSlidingPagination'
@@ -44,21 +49,16 @@ import DpTooltipIcon from './DpTooltipIcon/DpTooltipIcon'
 import DpTreeList from './DpTreeList/DpTreeList'
 
 import {
-  dataTableSearch,
   DpAnonymizeText,
   DpBulkEditHeader,
   DpChangeStateAtDate,
-  DpColumnSelector,
   DpCopyPasteButton,
   DpDashboardTaskCard,
-  DpDataTable,
-  DpDataTableExtended,
   DpEditableList,
   DpEditor,
   DpNotifyContainer,
   DpNotifyMessage,
   DpObscure,
-  DpSelectPageItemCount,
   DpStickyElement,
   DpTableCard,
   DpTableCardListHeader,
@@ -70,6 +70,7 @@ import {
 } from './core/'
 
 export {
+  dataTableSearch,
   DpAccordion,
   DpButton,
   DpButtonRow,
@@ -78,7 +79,6 @@ export {
   DpInput,
   DpLabel,
   DpLoading,
-  dataTableSearch,
   DpAnonymizeText,
   DpAutocomplete,
   DpBulkEditHeader,
